requirements.txt
alembic==1.6.3
appdirs==1.4.4
attrs==21.2.0
autoflake==1.4
beautifulsoup4==4.9.3
black==21.5b1
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
coverage==5.5
flake8==3.9.2
greenlet==1.1.0
idna==2.10
iniconfig==1.1.1
invoke==1.5.0
isort==5.8.0
lxml==4.6.3
Mako==1.1.4
MarkupSafe==2.0.1
mccabe==0.6.1
mypy==0.812
mypy-extensions==0.4.3
packaging==20.9
pathspec==0.8.1
pluggy==0.13.1
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==2.4.7
pytest==6.2.4
pytest-cov==2.12.0
pytest-mock==3.6.1
python-dateutil==2.8.1
python-decouple==3.4
python-editor==1.0.4
regex==2021.4.4
requests==2.25.1
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.15
sqlalchemy-mixins==1.3
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
urllib3==1.26.4
